refactor(works): extract shared holdings-resolver lib (#2453/#2567 convergence)
Lift the pure holdings primitives out of the SHWEP matcher into
scripts/lib/holdings-resolver.mjs so the works-catalog / translation registry
imports them instead of reinventing a second, divergent matcher (coordinated
with the #2626 translation-prior-layer session on #2567).

Exports (pure, no DB/Gemini): translatedRatio, isCollected, COLLECTED_RX,
bestEdition (completeness-tiered + dedicated-over-collected), editionReadable,
editionVisible, and holdingStatus (work-level 'absent'|'held_unprocessed'|
'held_readable'). Two predicates kept deliberately separate: OWNERSHIP+PROCESSED
(holdingStatus, computed over the FULL catalog) vs the PUBLIC link gate
(editionVisible, applied on top) — collapsing them is what nearly re-acquired
Philo we own (visible-only blind spot) and what would link readers to hidden books.

shwep-cited-works.mjs now imports these; --gap-audit uses editionReadable/
editionVisible for its reader-facing cut. Behavior unchanged (emit reproduces
held 297 / acquire 167).
